refactor(whm): 모달 컴포넌트 views/components → components/whManager 로 이동
