+++ src/components/whManager/InventoryDetailModal.vue
