+++ src/views/whManager/AsnListView.vue
@@ -4,8 +4,8 @@ import { useRouter } from 'vue-router'
 import AppLayout from '@/components/layout/AppLayout.vue'
 import BaseTable from '@/components/common/BaseTable.vue'
 import { ROUTE_NAMES } from '@/constants'
-import AsnDetailModal from './components/AsnDetailModal.vue'
-import AsnMismatchModal from './components/AsnMismatchModal.vue'
+import AsnDetailModal from '@/components/whManager/AsnDetailModal.vue'
+import AsnMismatchModal from '@/components/whManager/AsnMismatchModal.vue'
 import { useUiStore } from '@/stores/ui'
 
 const router = useRouter()
